requirements.txt
Flask==2.3.3
pandas==2.0.3
reportlab==4.0.4
fpdf2==2.7.4
gunicorn==21.2.0
